revert mistake placement of CheckerConfig
