+++ driver/checking/block_gas_rate.go
@@ -22,7 +22,6 @@ import (
 
 	"github.com/0xsoniclabs/norma/driver"
 	"github.com/0xsoniclabs/norma/driver/monitoring"
-	"github.com/0xsoniclabs/norma/driver/parser"
 )
 
 const defaultCeiling float64 = math.MaxFloat64
@@ -46,7 +45,7 @@ type blockGasRateChecker struct {
 // If the config doesn't provide any replacement value, copy from the value of the original.
 // If the config is invalid, return error instead.
 // If the config is nil, return original checker.
-func (c *blockGasRateChecker) Configure(config parser.CheckerConfig) Checker {
+func (c *blockGasRateChecker) Configure(config CheckerConfig) Checker {
 	if config == nil {
 		return c
 	}
