+++ driver/checking/blocks_rolling.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/0xsoniclabs/norma/driver"
 	"github.com/0xsoniclabs/norma/driver/monitoring"
-	"github.com/0xsoniclabs/norma/driver/parser"
 )
 
 const defaultToleranceSamples int = 10
@@ -27,7 +26,7 @@ type blocksRollingChecker struct {
 // If the config doesn't provide any replacement value, copy from the value of the original.
 // If the config is invalid, return error instead.
 // If the config is nil, return original checker.
-func (c *blocksRollingChecker) Configure(config parser.CheckerConfig) Checker {
+func (c *blocksRollingChecker) Configure(config CheckerConfig) Checker {
 	if config == nil {
 		return c
 	}
