+++ driver/checking/checker_mock.go
@@ -12,7 +12,6 @@ package checking
 import (
 	reflect "reflect"
 
-	parser "github.com/0xsoniclabs/norma/driver/parser"
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -55,7 +54,7 @@ func (mr *MockCheckerMockRecorder) Check() *gomock.Call {
 }
 
 // Configure mocks base method.
-func (m *MockChecker) Configure(arg0 parser.CheckerConfig) Checker {
+func (m *MockChecker) Configure(arg0 CheckerConfig) Checker {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Configure", arg0)
 	ret0, _ := ret[0].(Checker)
